Document assignation between changes

+++ app/models/client.rb
@@ -6,12 +6,16 @@ class Client < ActiveRecord::Base
   GENDER = [['Hombre', true], ['Mujer', false]]
   MARITAL_STATUS = [['Soltero', 'single'], ['Casado', 'married'], ['Viudo', 'widowed'], ['Divorciado', 'divorced']]
   CREDIT_TYPE = [['Credito Bancaro', 'bank'], ['Credito Infonavit', 'infonavit'], ['Credito FOVISSSTE', 'fovissste']]
-  FISCAL_ENTITY = [['Persona Asalariada', 'wage'], ['Persona Fisica', 'naural'], ['Persona Moral', 'legal']]
+  FISCAL_ENTITY = [['Persona Asalariada', 'wage'], ['Persona Fisica', 'natural'], ['Persona Moral', 'legal']]
   belongs_to :salesman, class_name: "User",   foreign_key: "current_salesman_id"
   has_many :appointments, class_name: "Appointment", dependent: :destroy
   has_many :interactions, dependent: :destroy
   has_many :phones, dependent: :destroy
   has_one :dossier, dependent: :destroy
+
+  #validates :marital_status, inclusion: ['single', 'married', 'widowed', 'divorced']
+  #validates :credit_type, inclusion: ['bank', 'infonavit', 'fovissste']
+  #validates :fiscal_entity, inclusion: ['wage', 'natural', 'legal']
 
   accepts_nested_attributes_for :phones, reject_if: :all_blank, allow_destroy: true
 
@@ -49,18 +53,29 @@ class Client < ActiveRecord::Base
     marital_status == 'married'
   end
 
+  def spouse_active
+    married? && spouse_works && credit_type == 'bank'
+  end
+
   def documentize
     if dossier.nil?
       create_dossier
       dossier.create_location_information
       dossier.create_nominal_work_record
+      dossier.make_checklist(credit_type, fiscal_entity, spouse_active)
     end
-    if married? && spouse_works
-      dossier.create_spouse_work_record
-    else
-      dossier.spouse_work_record.destroy if dossier.spouse_work_record
+    dossier.create_spouse_work_record if married? && spouse_works
+  end
+
+  def update_documents
+    if dossier != nil
+      dossier.make_checklist(credit_type, fiscal_entity, spouse_active)
+      if married? && spouse_works
+        dossier.create_spouse_work_record if dossier.spouse_work_record.nil?
+      else
+        dossier.spouse_work_record.destroy if dossier.spouse_work_record
+      end
     end
-      
   end
 
 private
